1.3.Form Type Class{Creating a Form Class}
Perfect! Back in our editor, once PhpStorm finishes indexing, we should be able to find the AbstractType class from the Form component. Got it! Now, go to the Code -> generate menu, or Cmd+N on a Mac, and click override methods. There are several methods that you can override to control different parts of your form. But, by far, the most important is buildForm().
Inside this method, our job is pretty simple: use this  object to, um... build the form! Use $builder->add() to add two fields right now: title and content. These are the two most important fields inside the Article entity class.
And... that's it! We'll do more work here later, but this is enough.