3.3.Success (Flash) Messages{Rendering the Flash Message}
Rendering the Flash Message
Oh, and the success key? I just made that up. That's sort of a "category" or "type", and we'll use it to read the message and render it. And... where should we read and render the message? The best place is in your base.html.twig layout. Why? Because no matter what page you redirect to after a form submit, your flash message will then be rendered.
Scroll down a little bit and find the block body. Right before this - so that it's not overridden by our child templates, add {% for message in app.flashes() %} and pass this our type: success. Remember: Symfony adds one global variable to Twig called app, which comes in handy here. Inside the for, add a div with class="alert alert-success" and, inside, print message.
Done! Oh, but, why do we need a for loop here to read the message? Well, it's not too common, but you can technically put as many messages onto your success flash type as you want. So, in theory, there could be 5 success messages that we need to read and print... but you'll usually have just one.
Anyways, let's try this crazy thang! Move back so we can create another important article:

  Ursa Minor: Major Construction Planned

Hit enter and... hello nice message! I don't like that weird margin issue - but we'll fix that in a minute. When you refresh, yep! The message disappears in a flash... because it was removed from the session when we read it the first time.