3.4.Success (Flash) Messages{Peeking at the Flash Messages}
Peeking at the Flash Messages
Ok, let's fix that ugly margin issue... it's actually interesting. Inspect element on the page and find the navbar. Ah, it has some bottom margin thanks to this mb-5 class. Hmm. To make this look right, we don't want to render that mb-5 class when there is a flash message. How can we do that?
Back in base.html.twig, scroll up a bit to find the navbar. Ok: we could count the number of success flash messages, and if there are more than 0, do not print the mb-5 class. That's pretty simple, except for one huge problem! If we read the flash messages here to count them, that would also remove them! Our loop below would never do anything!
How can we work around that? By peeking at the flash messages. Copy the class. Then, say app.session.flashbag.peek('success'). Pipe that to the length filter and if this is greater than zero, print nothing. Otherwise, print the mb-5 class.
This... deserves some explanation. First, the global app variable is actually an object called, conveniently, AppVariable! Press Shift+Shift and search for this so we can see exactly what it looks like.
Before, we used the getFlashes() method, which handles all the details of working with the Session object. But, if we need to "peek", we need to work with the Session directly via the getSession() shortcut. It turns out, the "flash messages" are stored on a sub-object called the "flash bag". This new longer code fetches the Session, gets that "FlashBag" and calls peek() on it.
Ok, let's see if that fixed things! Move back over and click to author another amazing article:

  Mars: God of War? Or Misunderstood?

Hit enter to submit and... got it! Flash message and no extra margin.
Next, let's learn how we can do... less work! By bossing around the form system and forcing it to create and populate our Article object so we don't have to.