7.3.EntityType: Drop-downs from the Database{Hello EntityType}
But, as soon as you pass the field type explicitly, form's guessyng system stops guessing anything. That means that we need to manually set class to User::class. This is why I often omit the 2nd argument if it's being guessed correctly. And, we could do that here.
Try it again. Got it!