7.5.EntityType: Drop-downs from the Database{Controlling the Option Display Value}
Want to get fancier? I thought you would. You can also pass this option a callback, which Symfony will call for each item and pass it the data for that option - a User object in this case. Inside, we can return whatever we want. How about return sprintf('(%d) %s') passing ->getId() and ->getEmail().
Cool! Refresh that! Got it!