10.1.Validation Constraints with @Assert{Adding your First Assert Annotation; The Built-in Validation Constraints}
When you talk about validation, what you're really talking about is business rules validation. That's where you tell Symfony that the title is required and needs to be a certain length, or that some field should be a valid email address, or that the password must contain 2 upper case letters, 2 lower case letters, 3 unicode characters and at least 4 emojis. It's about making the data constrain to your application's rules.

Adding your First Assert Annotation
Symfony's validation is kinda interesting because you do not apply the validation rules to the form. Nope, you apply them to your class via annotations. Check this out: I want the title field to be required. To do that, type @NotBlank and hit tab to autocomplete to @Assert\NotBlank. Because I have the PHP annotations plugin installed, when I auto-completed that, it added a use statement on top that we need: use Symfony\Component\Validator\Constraints as Assert.
Without doing anything else, refresh the form - the title field is empty. Yes! That's our error!

  This value should not be blank.

To customize that, add a message key to the annotation:

  Get creative and think of a title!
Try it again - refresh and... nice!

The Built-in Validation Constraints
On the docs, click to go back to the documentation homepage. Then, under guides, find the "Validation" guide. Just like with the form fields, there are a bunch of built-in validation constraints that... can help you validate almost anything! And... just like with form field types, each validation constraint has different options that control its behavior.
For example - check out Length: you can set the min length with the min option, or max with max. Control their error messages with minMessage and maxMessage.
Oh, another way to see what the options are is to remember that every annotation has a concrete PHP class behind it. Thanks to the PHP annotations plugin, I can hold Command or Ctrl and click the annotation to jump to that class.
Nice! Every property becomes an option that you can pass to the annotation. We'll see this again later when we create our own custom validation constraint.
Anyways, we won't talk too much about validation constraints because... they're honestly pretty easy: it's usually a matter of finding which validation constraint you need and the options to pass to it.