10.2.Validation Constraints with @Assert{The Callback Constraint}
The Callback Constraint
Oh, but there is one really cool constraint called Callback. This is the tool when you need to go rogue and do something totally custom. Check it out: create a method in your class and add @Assert\Callback() above it. Then, during validation, Symfony will call your method!
Let's copy this, find our Article class, go all the way to the bottom, and paste. Oh, I need to retype the end of ExecutionContextInterface and auto-complete it to get the use statement. Then, inside... it's awesome! We can do whatever we want!
Let's make sure that the title of this Article doesn't contain the string the borg... cause they're scary. So, if stripos() of $this->getTitle() and the borg does not equal false... error! To create the error, use $context->buildViolation():

  Um.. the Borg kinda makes us nervous

Apparently so nervous that I typed "the Bork" instead! Resistance to typos is futile...
Next, choose which field to attach the error to with ->atPath('title') and finish with ->addViolation(). That's it!
Go back to our form, write an article about how you really want to join the borg and Create!
Got it! Custom validation logic with a custom error.
Next: let's talk a little more about how we can control the rendering of these fields. Because, right now, we're just sort of rendering them all at once... without much control over their look and feel.