12.4.Form Rendering Variables{form_label(), form_widget(), form_help(), form _errors()}
When you render things at this level, you start to lose some of the special formatting that form_row() gives you. Sure, it's easy to re-add that div. But form_row also adds a special error class to that div when the field has a validation error.
For that reason, let's go back to using form_row().
A little bit later, we're going to learn how we can use form_row(), but completely customize how it looks for one specific form, or across your entire site!
We'll do this by creating a "form theme". It's kind of the best of both worlds: you can render things in the lazy way, but still have the control you need.
But before we get there - let's learn how to create an "edit" form so we can update articles!