13.4.The Edit Form{Tweaks for the Edit Form}
Next, even though it worked, we don't really want to re-use the same Twig template, because it has text like "Launch a new article" and "Create". Change the template name to edit.html.twig. Then, down in the templates/article_admin directory, copy the new.html.twig and name it edit.html.twig, because, there's not m
uch that needs to be different.
Update the h1 to Edit the Article and, for the button, Update.
Cool! Let's try this - refresh! Looks perfect! Let's change some content, hit Update and... we're back!