14.3.Registration Form{Making firstName Optional}
Making firstName Optional
Before we test it, open your User entity class. We originally made the firstName field not nullable. That's the default value for nullable. So if you don't see nullable=true, it means that the field is required in the database.
Now, I do want to allow users to register without their firstName. No problem: set nullable=true.
Then, find your terminal and run:

php bin/console make:migration
Let's go check out that new file. Yep! No surprises: it just makes the column not required.
Move back over and run this with:

php bin/console doctrine:migrations:migrate
Excellent! Let's try to register! Register as geordi@theenterprise.org, password, of course, engage. Hit enter and... nice! We are even logged in as Geordi!
Next: we have a problem! We're temporarily storing the plaintext password on the password field... which is a big no no! If something goes wrong, we might accidentally save the user's plaintext password to the database.
To fix that, we, for the first time, will add a field to our form that does not exist on our entity. An awesome feature called mapped will let us do that.